Add content script such that a blue square appears on bottom right of every page when DIH is on (placeholder for DIH)
